TS: DomainsOnProtein [WIP]
