+++ src/components/Protein/Isoforms/Viewer/index.js
@@ -8,6 +8,7 @@ import loadData from 'higherOrder/loadData';
 import descriptionToPath from 'utils/processDescription/descriptionToPath';
 
 import NumberComponent from 'components/NumberComponent';
+// $FlowFixMe
 import { groupByEntryType } from 'components/Related/DomainsOnProtein';
 // $FlowFixMe
 import { byEntryType } from 'components/Related/DomainsOnProtein/DomainsOnProteinLoaded';
